+++ hendricks-bot/src/CommandRegistry.ts
@@ -1,0 +1,9 @@
+import Command from './types/Command';
+
+export default class CommandRegistry {
+	static commands: Command[] = [];
+
+	static register(...c: Command[]) {
+		this.commands.concat(c);
+	}
+}
