--- hendricks-bot/src/deployCommands.ts
@@ -1,32 +1,0 @@
-import { REST, Routes } from 'discord.js';
-import { config } from 'dotenv';
-import commands from './commands';
-import logger from './logger';
-
-config();
-const token = process.env['TOKEN'] as string;
-const guildId = process.env['GUILD_ID'] as string;
-const clientId = process.env['CLIENT_ID'] as string;
-
-const rest = new REST({ version: '10' }).setToken(token);
-
-// add guild only commands
-rest.put(Routes.applicationGuildCommands(clientId, guildId), { body: [] })
-	.then((data) =>
-		logger.info(
-			`Successfully registered ${(data as any[]).length} guild commands.`
-		)
-	)
-	.catch(console.error);
-
-// add global commands
-const applicationCommands = commands.map((command) => command.slash.toJSON());
-rest.put(Routes.applicationCommands(clientId), { body: applicationCommands })
-	.then((data) =>
-		logger.info(
-			`Successfully registered ${
-				(data as any[]).length
-			} application commands.`
-		)
-	)
-	.catch(console.error);
