+++ hendricks-bot/src/events/index.ts
@@ -1,7 +1,4 @@
 import ready from './ready';
-import interactionCreate from './interactionCreate';
-// import messageCreate from './messageCreate';
-import messageReactionAdd from './messageReactionAdd';
 
 // export default [ready, interactionCreate, messageCreate, messageReactionAdd];
-export default [ready, interactionCreate, messageReactionAdd];
+export default [ready];
