+++ hendricks-bot/src/plugins/codeblocks/commands/index.ts
@@ -1,0 +1,3 @@
+import eval from './eval';
+
+export default [eval];
