+++ hendricks-bot/src/plugins/codeblocks/index.ts
@@ -1,0 +1,13 @@
+import Plugin from '@hendricks/types/Plugin';
+import codeblockReactions from './codeblockReactions';
+import codeblockRunner from './codeblockRunner';
+import commands from './commands';
+
+export default {
+	id: 'codeblock-reactions',
+	start() {
+		codeblockReactions();
+		codeblockRunner();
+	},
+	commands,
+} as Plugin;
