+++ hendricks-bot/src/plugins/plugins.json
@@ -1,3 +1,4 @@
 [
-	"codeblock-reactions"
+	"codeblocks",
+	"stable-diffusion"
 ]
