+++ hendricks-bot/src/plugins/stable-diffusion/commands/index.ts
@@ -1,0 +1,4 @@
+import imagine from './imagine';
+import queued from './queued';
+
+export default [imagine, queued];
