+++ hendricks-bot/src/plugins/stable-diffusion/tasks.ts
