+++ hendricks-bot/src/types/Plugin.d.ts
@@ -1,5 +1,8 @@
+import Command from './Command';
+
 export default interface Plugin {
 	id: string;
 	init?(): Promise<void>;
-	start(): void;
+	start?(): void;
+	commands?: Command[];
 }
