+++ hendricks-bot/src/util/commandUtils.ts
